Separate an error module
